requirements.txt
flask
python-dotenv
gunicorn
psycopg2-binary
requests
pytz
